fix: 매크로 | 매크로 에러 처리
Former-commit-id: 720991f1f0fc8abdf66d9b12f3b947b86991c5d7

+++ AOS/Sonmal/app/src/main/java/com/d202/sonmal/ui/macro/MacroAddFragment.kt
@@ -204,11 +204,16 @@ class MacroAddFragment: Fragment() {
 
     private fun initObserver() {
         macroViewmodel.macroAddCallback.observe(viewLifecycleOwner) {
-            Toast.makeText(requireContext(), "등록 완료", Toast.LENGTH_SHORT).show()
-            if(videoFileSave != null) {
-                videoFileSave!!.delete()
+            if(it == 200) {
+                Toast.makeText(requireContext(), "등록 완료", Toast.LENGTH_SHORT).show()
+                if(videoFileSave != null) {
+                    videoFileSave!!.delete()
+                }
+                findNavController().navigateUp()
+            } else {
+                Toast.makeText(requireContext(), "등록 실패", Toast.LENGTH_SHORT).show()
             }
-            findNavController().navigateUp()
+
 
         }
         macroViewmodel.refreshExpire.observe(viewLifecycleOwner) {
